requirements.txt
google-generativeai==0.5.4
langchain==0.1.0
streamlit==1.27.2
